Fixed add, and hidden brokenness in delete
Add was not working because I forgot to ask for the expiration date.
Delete was not working because I was not deleting the rows from the
tables in the reverse order that they were added. Add inserted into
credit_card, then billing_info. Delete SHOULD have been deleting from
billing_info THEN credit_card. It works like a charm now.